Fix blur when image is not loaded (see desc)
Previously, the OctoImage was always the child of the ColorFiltered and the ImageFiltered. This caused the placeholders to be filtered, causing a black square in the corner of the screen when loaded. Now, the filters are applied in the OctoImage's imageBuilder. I've also removed the AnimatedSwitcher because it didn't make sense with the main album art not fading.