Index image downloads by blurhash instead of ID
This stops Finamp from downloading identical images. Blurhashes are used to test this as it doesn't require downloading the images to check if they're identical.

Note that this commit does not include a migration from the old ID-based system, so weird things will happen if you update Finamp to this commit.