:bug: Remove the playback state age check again for now
The playbackEventStream unfortunately isn't frequent enough, causing the previous event to sometimes be older than the specified maximum age.
This causes stop events to be dropped sometimes.
Thus, we revert this, but also ensure that the position, that is automatically extrapolated based on the current time, doesn't exceed the duration of the track.