Add Initial Certificate Override Support
Backend work to provide certificate override functionality added. User Interface to manage history and provide context should be added. Clarification of when new certs for a domain (if not valid) are added versus never having been approved should be noted to the user. Allow them to reject if the remote end has changed (or is behind a load-balancer with multiple certs/ports...)

Signed-off-by: Odin Vex <44311901+OdinVex@users.noreply.github.com>