:bug: Fix sort by DateCreated not working when offline
By default, the DateCreated field is missing from the Jellyfin Items response.
This causes the DateCreated field to always be null on BaseItemDtos,
so it's not available for sorting when offline.

By adding the field name to the defaultFields, this is resolved.
However, all items will have to be resynced before sorting works offline.