Make sure audio handler knows queue from the start
If we swipe to add to the queue after launching the app (when no queue exists), the new audio source will temporarily be added to the _queueAudioSource, but since the _audioHandler doesn't know this queue yet, _queueFromConcatenatingAudioSource will still see an empty queue, and thus ignore the new element. Immediately registering the audio source to the audio handler resolves this issue.
Furthermore, since the two elements are only written once, they can be final.