Fix bottom sheets on foldables
Previously, bottom sheets would snap to the center of the screen on (Flip-style) foldable devices when in half opened mode.
To fix that, an anchor point has to be supplied to the showModalBottomSheet function.

However, when in half opened mode, the queue bottom sheet can only fill its assigned subscreen - in our case, the lower half of the screen. Currently, it'll only fill 92% of that subscreen, which looks and feels awkward. Thus, we check whether the device is currently half-opened, and set it to 100% of the subscreen (or 50% of the screen) in that case.
Unfortunately, we cannot use the BuildContext of the modal bottom sheet for that, since it only exposes display features for the current subscreen, and doesn't actively react to changes. To resolve that, we use a mix of WidgetsBindingObserver, ChangeNotifier(s), Providers and ListenableBuilders.
For other bottom sheets, we keep their current size.