queu_service: remove initializePlayer
initializePlayer was causing the audio source to be set twice during
startup, because it's also set after queue restore.

On Android, when the app is restarted from the play button on the media
notification, it gets a play request before the queue is restored.  If
the player is already initialised with an empty queue (by
initializePlayer) it will immediately stop, and won't restart after the
queue is restored.