Add support for Android user-installed certificates
This fixes https://github.com/jmshrv/finamp/issues/106

Frustratingly, Dart/Flutter ignores user-installed certificates. Working
around this requires rooting your Android device to install certificates
as "system" certs, which isn't an option for all people.

This is a known issue with Dart, see
https://github.com/dart-lang/sdk/issues/50435 and
https://github.com/flutter/flutter/issues/56607 for details.

This depends on my fork of
`johnstef99/flutter_user_certificates_android`, which I've sent a PR for
here
<https://github.com/johnstef99/flutter_user_certificates_android/pull/2>.
If y'all don't like the supply chain implications of that, I'm happy to
inline the implementation here instead.

I opted not to make this configurable as I don't see the value. People
using finamp are sophisticated self-hosters, and should understand the
risks of configuring their OS to trust certificates.