Fix files being marked as cache for older downloads on Android
An earlier version of the background_downloader library contained a bug
that caused all downloaded files to be marked as cache files on Android.
This was due to an incorrect group id being set on those files.
The newest version fixes this issue, but we have to update all
previously downloaded files to the correct group id.